Make it work with Falcon's default SSL server
Falcon runs the server on https://localhost:9292 by default, using a
self-signed certificate. In order for Shrine::Storage::Tus to be able to
download the uploaded tus file during Shrine promotion, its downloader
needs to be configured to use Falcon's self-signed certificate. We can
fetch it with the localhost gem.